Rename the NEON and add SSE4.2-based JSON SIMD implementations with `daw::simd` abstraction, modularize SIMD headers, simplify parsing logic, and update DAW_JSON_HAS_SIMD macro usage.
