+++ include/daw/json/impl/daw_json_parse_policy.h
@@ -19,7 +19,6 @@
 #include "daw/json/impl/daw_json_parse_policy_no_comments.h"
 #include "daw/json/impl/daw_json_parse_policy_policy_details.h"
 #include "daw/json/impl/daw_json_skip_class_simd.h"
-#include "daw/json/impl/daw_json_skip_class_sse42.h"
 #include "daw/json/impl/daw_json_string_util.h"
 
 #include <daw/cpp_17.h>
@@ -524,13 +523,6 @@ namespace daw::json {
 						  json_details::SkipBracketedType::Class>( *this );
 					}
 				}
-#elif defined( DAW_ALLOW_SSE42 )
-				if constexpr( std::is_same_v<CommentPolicy, NoCommentSkippingPolicy> ) {
-					if( not DAW_IS_CONSTANT_EVALUATED( ) ) {
-						return json_details::skip_bracketed_item_sse42<
-						  json_details::SkipBracketedType::Class>( *this );
-					}
-				}
 #endif
 				  if constexpr( is_unchecked_input ) {
 					return skip_bracketed_item_unchecked<
@@ -551,13 +543,6 @@ namespace daw::json {
 				if constexpr( std::is_same_v<CommentPolicy, NoCommentSkippingPolicy> ) {
 					if( not DAW_IS_CONSTANT_EVALUATED( ) ) {
 						return json_details::skip_bracketed_item_simd<
-						  json_details::SkipBracketedType::Array>( *this );
-					}
-				}
-#elif defined( DAW_ALLOW_SSE42 )
-				if constexpr( std::is_same_v<CommentPolicy, NoCommentSkippingPolicy> ) {
-					if( not DAW_IS_CONSTANT_EVALUATED( ) ) {
-						return json_details::skip_bracketed_item_sse42<
 						  json_details::SkipBracketedType::Array>( *this );
 					}
 				}
