+++ include/daw/json/impl/daw_json_simd.h
@@ -20,128 +20,12 @@
 
 #endif
 
-#if ( defined( __ARM_NEON ) or defined( __ARM_NEON__ ) ) and \
-  DAW_CPP_VERSION >= 202002L
+#if ( defined( __ARM_NEON ) or defined( __ARM_NEON__ ) )
+#include "daw/json/impl/daw_json_simd_neon.h"
+#endif
 
-#define DAW_JSON_HAS_NEON_SIMD 1
-
-#include <arm_neon.h>
-
-#include <cstddef>
-#include <cstdint>
-#include <cstring>
-#include <type_traits>
-
-namespace daw::simd_impl::neon {
-	namespace details {
-		[[nodiscard]] inline std::uint64_t
-		to_bits( uint8x16_t comparison ) noexcept {
-			alignas( 16 ) static constexpr std::uint8_t bit_values[16] = {
-			  0x01, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x80,
-			  0x01, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x80 };
-			auto const bits = vandq_u8( comparison, vld1q_u8( bit_values ) );
-			auto const pairs = vpaddlq_u8( bits );
-			auto const quads = vpaddlq_u16( pairs );
-			auto const octets = vpaddlq_u32( quads );
-			return vgetq_lane_u64( octets, 0 ) |
-			       ( vgetq_lane_u64( octets, 1 ) << 8U );
-		}
-	} // namespace details
-
-	class mask {
-		uint8x16_t m_value;
-
-	public:
-		explicit mask( uint8x16_t value ) noexcept
-		  : m_value( value ) {}
-
-		[[nodiscard]] std::uint64_t to_ullong( ) const noexcept {
-			return details::to_bits( m_value );
-		}
-
-		[[nodiscard]] friend mask operator|( mask lhs, mask rhs ) noexcept {
-			return mask( vorrq_u8( lhs.m_value, rhs.m_value ) );
-		}
-
-		[[nodiscard]] friend mask operator&( mask lhs, mask rhs ) noexcept {
-			return mask( vandq_u8( lhs.m_value, rhs.m_value ) );
-		}
-
-		[[nodiscard]] friend mask operator!( mask value ) noexcept {
-			return mask( vmvnq_u8( value.m_value ) );
-		}
-	};
-
-	template<typename T, std::size_t /*MaximumSize*/>
-	class vec {
-		static_assert( sizeof( T ) == 1,
-		               "The NEON JSON SIMD implementation supports byte types" );
-		static_assert( std::is_integral_v<T>,
-		               "The NEON JSON SIMD implementation requires an integral type" );
-
-		uint8x16_t m_value;
-
-		template<typename U, std::size_t MaximumSize>
-		friend class vec;
-
-		template<typename Simd, typename Range, typename Flag>
-		friend Simd unchecked_load( Range values, Flag );
-
-		template<typename Simd, typename Range, typename Flag>
-		friend Simd partial_load( Range values, Flag );
-
-		explicit vec( uint8x16_t value ) noexcept
-		  : m_value( value ) {}
-
-	public:
-		using value_type = T;
-		using mask_type = mask;
-
-		explicit vec( value_type value ) noexcept
-		  : m_value(
-		      vdupq_n_u8( static_cast<std::uint8_t>( value ) ) ) {}
-
-		[[nodiscard]] static constexpr std::size_t size( ) noexcept {
-			return 16;
-		}
-
-		[[nodiscard]] friend mask operator==( vec lhs, vec rhs ) noexcept {
-			return mask( vceqq_u8( lhs.m_value, rhs.m_value ) );
-		}
-
-		[[nodiscard]] friend mask operator>=( vec lhs, vec rhs ) noexcept {
-			return mask( vcgeq_u8( lhs.m_value, rhs.m_value ) );
-		}
-
-		[[nodiscard]] friend mask operator<=( vec lhs, vec rhs ) noexcept {
-			return mask( vcleq_u8( lhs.m_value, rhs.m_value ) );
-		}
-	};
-
-	struct flag_default_t {};
-	struct flag_convert_t {};
-
-	inline constexpr flag_default_t flag_default = {};
-	inline constexpr flag_convert_t flag_convert = {};
-
-	template<typename Simd, typename Range, typename Flag = flag_default_t>
-	[[nodiscard]] inline Simd unchecked_load(
-	  Range values, Flag = flag_default ) {
-		static_assert( Simd::size( ) == 16 );
-		return Simd( vld1q_u8(
-		  reinterpret_cast<std::uint8_t const *>( values.data( ) ) ) );
-	}
-
-	template<typename Simd, typename Range, typename Flag = flag_default_t>
-	[[nodiscard]] inline Simd partial_load(
-	  Range values, Flag = flag_default ) {
-		static_assert( Simd::size( ) == 16 );
-		auto result = vdupq_n_u8( 0 );
-		std::memcpy( &result, values.data( ), values.size( ) );
-		return Simd( result );
-	}
-} // namespace daw::simd_impl::neon
-
+#if defined( __SSE4_2__ ) or defined( __AVX__ ) or defined( __AVX2__ )
+#include "daw/json/impl/daw_json_simd_sse42.h"
 #endif
 
 #if defined( DAW_JSON_HAS_STD_SIMD )
@@ -160,6 +44,15 @@ namespace daw {
 
 namespace daw {
 	namespace simd = simd_impl::neon;
+}
+
+#elif defined( DAW_JSON_HAS_SSE42_SIMD )
+
+#define DAW_JSON_HAS_SIMD 1
+#define DAW_JSON_SIMD_CONSTEXPR inline
+
+namespace daw {
+	namespace simd = simd_impl::sse42;
 }
 
 #endif
